More expr analysis work. Fix expr formatting memoization of filtered tables per #668
